requirements.txt
# Lab 7 CDDA - Python Dependencies
# Install with: pip install -r requirements.txt

# Azure SDKs
azure-storage-blob>=12.19.0
azure-storage-queue>=12.9.0
azure-cosmos>=4.5.1
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-eventhub>=5.11.0
azure-monitor-opentelemetry>=1.0.0
applicationinsights>=0.11.10

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
joblib>=1.3.0

# MLflow
mlflow>=2.9.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# UI & Visualization
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0

